Removed 'using namespace std;' from all header files.
Remark: Please make sure NEVER to import a namespace (or part of a namespace) at GLOBAL SCOPE of a header file!
If you do this in implementation (.cc) files, it is (more or less) ok, because it only affects this source module. To use it in header files causes everyone who imports your header file to use this 'using' directive, probably without him knowing. At best, this will result in hard to trace down compile-time errors; however, it may cause subtle bugs to pass unnoticed (e.g. if you forgot to implement a method named the same as one of the methods in the 'imported' namespace).
Note that it is perfectly ok to import namespaces (or parts of them) at a LOCAL SCOPE (as for example inside a method).