Added dummy WorldModifier. Implemented handling of WorldModifierEvents.
The TestWorldModifier class is a first concrete implementation of a WorldModifier. The synchronous ASG (with WMs) currently adds this modifier to every WorldModifierEvent (to be replaced with a flexible selection of modifiers). The EventHandler has been extended so that it processes all modifiers included in a WorldModifierEvent. This includes the processing of an anonymous HandleRequestEvent (all requests are handled within the same WorldModifierEvent).