TestWorldModifier now creates PositionRequests. Removed View from WorldModifier class.
The TestWorldModifier class now generates PositionRequests (not fully thought through). However, in general the modifier does what it is supposed to do: modify the world. The GlobalView has been removed, since it could not be used as planned (Views want to know which Robot is looking, WorldModifiers are no Robots). Instead, the WorldModifier directly accesses the WorldInformation class.