Added support for MarkerInformation objects to the WorldInformation wrapper.
The wrapper class now provides the Lua scripts with the ability to request MarkerInformation objects for a given id. The code still needs checks (i.e. whether an id exists at all, there are corresponding TODO comments). However, there seems to be a bigger problem: Parts of the existing bindings, on which my changes are built upon, do not work.