Lua bindings for Statistics moduled, example Lua script.
The bindings now provide access to the StatsCalc class. A Lua script has been added to the test project to demonstrate how a simple statistics script can be implemented. The scripts uses the methods to calculate defects every 100 rounds and writes the results to a separate file. This file is identified by an ISO timestampt (just take a look at CalculateDefects.lua, the code should be self-explanatory).